requirements.txt
Django==1.9.4
ExifRead==2.1.2
IPTCInfo==1.9.5.post6
Paste==2.0.3
Pillow==3.1.1
python-xmp-toolkit==2.0.1
pytz==2015.7
six==1.10.0
